--- src/gameMockups/Main.java
@@ -1,41 +1,0 @@
-package gameMockups;
-
-import java.awt.EventQueue;
-
-import javax.swing.UIManager;
-
-public class Main {
-	public static void main(String[] args) {
-		EventQueue.invokeLater(new Runnable() {
-			public void run() {
-				try {
-					UIManager.setLookAndFeel(UIManager.getSystemLookAndFeelClassName());
-					SelectLevel frame = new SelectLevel();
-					frame.setVisible(true);
-				} catch (Exception e) {
-					e.printStackTrace();
-				}
-			}
-		});
-	}
-}
-
-
-/**
- * Launch the application.
- */
-
-/*public class Main {
-	public static void main(String[] args) {
-		EventQueue.invokeLater(new Runnable() {
-			public void run() {
-				try {
-					SplashScreen frame = new SplashScreen();
-					frame.setVisible(true);
-				} catch (Exception e) {
-					e.printStackTrace();
-				}
-			}
-		});
-	}
-}*/
