+++ src/view/BoardView.java
@@ -1,4 +1,4 @@
-package builderMockups;
+package view;
 
 import java.awt.Dimension;
 import java.awt.event.MouseAdapter;
