+++ src/view/BuilderView.java
@@ -1,4 +1,4 @@
-package builderMockups;
+package view;
 
 import javax.swing.JFrame;
 import javax.swing.JPanel;
