+++ src/view/LevelInfoView.java
@@ -1,4 +1,4 @@
-package gameMockups;
+package view;
 
 import javax.swing.JPanel;
 import javax.swing.GroupLayout;
