+++ src/view/LevelPropertiesView.java
@@ -1,4 +1,4 @@
-package builderMockups;
+package view;
 
 import java.awt.Dimension;
 import java.awt.Font;
