+++ src/view/PlayingPieceGroupView.java
@@ -1,4 +1,4 @@
-package builderMockups;
+package view;
 
 import javax.swing.GroupLayout;
 import javax.swing.GroupLayout.Alignment;
