+++ src/view/SelectLevel.java
@@ -1,4 +1,4 @@
-package gameMockups;
+package view;
 import java.awt.BorderLayout;
 import java.awt.EventQueue;
